+++ samples/MLFlow.Sample.CLI/appsettings.json
@@ -1,0 +1,6 @@
+{
+    "MLFlowConfiguration": {
+        "MLFlowServerBaseUrl": "http://localhost:5000",
+        "APIBase": "api/2.0/preview/mlflow/" 
+    } 
+}
