+++ src/MLFlow.NET/MLFlow.NET.sln
@@ -7,9 +7,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MLFlow.NET.Lib", "MLFlow.NE
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MLFlow.NET.Tests", "MLFlow.NET.Tests\MLFlow.NET.Tests.csproj", "{D7E75D6B-239F-4764-AC59-ED9C7783B9A3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MLFlow.Sample.CLI", "MLFlow.Sample.CLI\MLFlow.Sample.CLI.csproj", "{26A998D0-CCD5-4BAF-B165-44F63C2ED73F}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{471D3F2B-628C-495B-889B-32A7EBDDD6D4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApplication1", "..\..\samples\web\WebApplication1\WebApplication1.csproj", "{F818F83F-555A-44C7-9699-C56B81639FE2}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MLFlow.Sample.CLI", "..\..\samples\MLFlow.Sample.CLI\MLFlow.Sample.CLI.csproj", "{F8406635-D1B6-40EC-8E87-8211C89455CD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,16 +27,21 @@ Global
 		{D7E75D6B-239F-4764-AC59-ED9C7783B9A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D7E75D6B-239F-4764-AC59-ED9C7783B9A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D7E75D6B-239F-4764-AC59-ED9C7783B9A3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{26A998D0-CCD5-4BAF-B165-44F63C2ED73F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{26A998D0-CCD5-4BAF-B165-44F63C2ED73F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{26A998D0-CCD5-4BAF-B165-44F63C2ED73F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{26A998D0-CCD5-4BAF-B165-44F63C2ED73F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F818F83F-555A-44C7-9699-C56B81639FE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F818F83F-555A-44C7-9699-C56B81639FE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F818F83F-555A-44C7-9699-C56B81639FE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F818F83F-555A-44C7-9699-C56B81639FE2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8406635-D1B6-40EC-8E87-8211C89455CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8406635-D1B6-40EC-8E87-8211C89455CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8406635-D1B6-40EC-8E87-8211C89455CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8406635-D1B6-40EC-8E87-8211C89455CD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{26A998D0-CCD5-4BAF-B165-44F63C2ED73F} = {471D3F2B-628C-495B-889B-32A7EBDDD6D4}
+		{F818F83F-555A-44C7-9699-C56B81639FE2} = {471D3F2B-628C-495B-889B-32A7EBDDD6D4}
+		{F8406635-D1B6-40EC-8E87-8211C89455CD} = {471D3F2B-628C-495B-889B-32A7EBDDD6D4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2C0816C7-D434-41D2-8DB3-E4BE6E43ECB7}
