--- src/MLFlow.NET/MLFlow.Sample.CLI/appsettings.json
@@ -1,6 +1,0 @@
-{
-    "MLFlowConfiguration": {
-        "MLFlowServerBaseUrl": "http://localhost:5000",
-        "APIBase": "api/2.0/preview/mlflow/" 
-    } 
-}
